python/requirements.txt
Click==7.0
grpcio==1.20.0
grpcio-tools==1.20.0
pkg-resources==0.0.0
protobuf==3.7.1
six==1.12.0
